01_02_graphql_bad/requirements_01_02.txt
python-dotenv==1.0.1
fastapi==0.112.0
uvicorn==0.30.5
SQLAlchemy==1.4.53
psycopg2==2.9.9
graphene==2.1.9
graphene-sqlalchemy==2.3.0
starlette-graphql==0.1.0